fix: Implementado un toast que avisa al usuario que no puede acceder al chat de un Trail cuando no está logueado o no pertenece al Trail
