+++ src/i18n/languages/en_US.tsx
@@ -451,6 +451,7 @@ const Trail = {
     rating: 'Rating',
 
     // TrailButtons
+    toast_alert_chat: 'No puedes acceder al chat si no estás logueado o no no estás unido al Trail.',
     chat: 'Chat',
   },
